refactor: move async query execution code outside of client.py
