+++ posthog/clickhouse/client/__init__.py
@@ -1,0 +1,3 @@
+from posthog.clickhouse.client.execute_async import execute_with_progress
+
+__all__ = ["execute_with_progress"]
